Add a view to stream results from a Celery task
Simulate that TODOs are generated by a celery task.

The client issues a POST request to this view, which starts a
celery task to generate TODOs, returning a task `task_id`.

The client then issues GET requests to this view with the
task ID, plus the last sequence number received. The view
responds with the TODOs generated since the last one.

When the task is complete, the GET request will return an
`end` parameter, that the client can use to know when to stop.

If an error occurred during the task, an `error` field is
returned. If there are TODOs generated before the error since the
last time, they are returned and the HTTP status is OK. If there
are no TODOs generated since the last time, in case of an error
the HTTP status is Internal Server Error.